+++ R/py_require.R
@@ -636,7 +636,11 @@ uv_binary <- function(bootstrap_install = TRUE) {
     ## multiple uv installations attempt to modify that config file.
   }
 
-  if (bootstrap_install) {
+  if (!bootstrap_install) {
+    uv <- NULL
+    return()
+  }
+
     # Install 'uv' in the 'r-reticulate' sub-folder inside the user's cache directory
     # https://github.com/astral-sh/uv/blob/main/docs/configuration/installer.md
 
@@ -679,7 +683,6 @@ uv_binary <- function(bootstrap_install = TRUE) {
       })
 
     }
-  }
 
   # if we bootstrap-installed successfully, return the path to the uv binary
   # if not, reset `uv` for the on.exit() hook and return NULL visibly