+++ NEWS.md
@@ -1,3 +1,5 @@
+# reticulate (development version)
+
 # reticulate 1.44.1
 
 - The default Python version in `install_python()`