+++ NEWS.md
@@ -39,6 +39,9 @@
   `repl_python()`. Added support for assigning `%system` command output to multiple
   variables via unpacking (#1844).
 
+- reticulate now warns when `py_require()`d packages are not found in the selected
+  Python virtual environment. This behavior can be disabled by setting the environment variable `RETICULATE_CHECK_REQUIRED_PACKAGES=0` (#1850).
+
 # reticulate 1.43.0
 
 - Fixed usage of micromamba and mamba, next-generation conda environment management tools.