+++ NEWS.md
@@ -2,6 +2,10 @@
 
 - Fix spurious `Error: ignoring SIGPIPE signal` on Unix when embedded Python writes to a closed pipe (#1868).
 
+- Fixed Positron UI interactions, such as restart/create new session with reticulate (#1871, #1869).
+
+- Added support for F1 help for reticulate Python objects in Positron (#1866).
+
 # reticulate 1.44.1
 
 - The default Python version in `install_python()`